tikfakebot/requirements.txt
python-telegram-bot
TikTokApi~=3.9.5
